requirements.txt
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==2.0.2
django-appconf==1.0.1
-e git+https://github.com/yourlabs/django-autocomplete-light.git@8e92292af23c2fa6de0d04e6c09eb374ea59588d#egg=django_autocomplete_light
django-classy-tags==0.6.2
-e git://github.com/thienkimlove/django-constance.git@23d1737f1ef5d9e7266047ae4beb0482bb95d9b4#egg=django_constance
django-debug-toolbar==1.9.1
django-djconfig==0.6.0
django-etc==0.10.0
-e git://github.com/smacker/django-filebrowser-no-grappelli.git@60acccf93561019d5f9d65519b4a9ac3ed0f5298#egg=django_filebrowser_no_grappelli
django-import-export==1.0.0
django-js-asset==1.0.0
django-jsoneditor==0.0.12
django-picklefield==1.0.0
django-select2==6.0.2
django-siteprefs==0.7.0
django-taggit==0.22.2
docutils==0.14
et-xmlfile==1.0.1
fake-factory==9999.9.9
idna==2.6
jdcal==1.3
jsonfield==2.0.2
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.0
packaging==17.1
pafy==0.3.82
persisting-theory==0.2.1
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
